requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4
allure-pytest-bdd==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==2.0.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
cmd2==1.5.0
colorama==0.4.4
commonmark==0.9.1
cryptography==35.0.0
execnet==1.9.0
glob2==0.7
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.0
loguru==0.5.3
Mako==1.1.6
marker==2.1.4
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.5
outcome==1.1.0
packaging==21.3
pandas==1.3.5
parse==1.19.0
parse-type==0.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pytest==6.2.5
pytest-bdd==5.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.25.1
rich==10.2.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
wsproto==1.0.0
xdist==0.0.1
yarl==1.7.2
